+++ chains/nomad-ethereum/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Unreleased
 
+- impl `std::fmt::Display` for `EthereumHome` and `EthereumReplica`
 - use gelato-sdk as a github dep rather than a crate
 - adds a changelog