+++ nomad-test/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- implement `Display` for mock contracts
 - add helper fn for testing with an http mock response
 - add mockito for mocking http responses
 - adds a changelog